+++ apps/backoffice-v2/src/lib/blocks/variants/constants.tsx
@@ -1,15 +1,28 @@
-import { KybManualReviewLegacyBlocks } from '@/lib/blocks/variants/KybManualReviewLegacyBlocks/KybManualReviewLegacyBlocks';
+import { ManualReviewBlocks } from '@/lib/blocks/variants/ManualReviewBlocks/ManualReviewBlocks';
 import { KybDemoBlocks } from '@/lib/blocks/variants/KybDemoBlocks/KybDemoBlocks';
 import { DefaultBlocks } from '@/lib/blocks/variants/DefaultBlocks/DefaultBlocks';
+import { FunctionComponent } from 'react';
+import { TWorkflowById } from '@/domains/workflows/fetchers';
+import { WorkflowVariant } from '@ballerine/common';
 
-export const UiSchemaVariant = {
-  KYB_MANUAL_REVIEW_LEGACY: 'KYB_MANUAL_REVIEW_LEGACY',
-  KYB_DEMO: 'KYB_DEMO',
-  DEFAULT: 'DEFAULT',
-} as const;
+export const BlocksVariant: FunctionComponent<{
+  workflowDefinition: Pick<TWorkflowById['workflowDefinition'], 'variant' | 'config' | 'version'>;
+}> = ({ workflowDefinition }) => {
+  if (
+    workflowDefinition?.version >= 0 &&
+    workflowDefinition?.variant === WorkflowVariant.KYB &&
+    workflowDefinition?.config?.isDemo
+  ) {
+    return <KybDemoBlocks />;
+  }
 
-export const uiSchemaVariantToBlocksMap = {
-  [UiSchemaVariant.KYB_MANUAL_REVIEW_LEGACY]: KybManualReviewLegacyBlocks,
-  [UiSchemaVariant.KYB_DEMO]: KybDemoBlocks,
-  [UiSchemaVariant.DEFAULT]: DefaultBlocks,
-} as const;
+  if (
+    workflowDefinition?.version >= 0 &&
+    workflowDefinition?.variant === WorkflowVariant.MANUAL_REVIEW &&
+    workflowDefinition?.config?.isLegacyReject
+  ) {
+    return <ManualReviewBlocks />;
+  }
+
+  return <DefaultBlocks />;
+};