+++ apps/backoffice-v2/src/pages/Entity/hooks/useEntityLogic/useEntityLogic.tsx
@@ -1,4 +1,3 @@
-import { UiSchemaVariant, uiSchemaVariantToBlocksMap } from '@/lib/blocks/variants/constants';
 import { useParams } from 'react-router-dom';
 import { useFilterId } from '@/common/hooks/useFilterId/useFilterId';
 import { useWorkflowQuery } from '@/domains/workflows/hooks/queries/useWorkflowQuery/useWorkflowQuery';
@@ -8,11 +7,9 @@ export const useEntityLogic = () => {
   const filterId = useFilterId();
   const { data: workflow } = useWorkflowQuery({ workflowId: entityId, filterId });
   const selectedEntity = workflow?.entity;
-  const BlocksByUiSchemaVariant = uiSchemaVariantToBlocksMap[UiSchemaVariant.DEFAULT];
 
   return {
     selectedEntity,
     workflow,
-    BlocksByUiSchemaVariant,
   };
 };