+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.4.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.35
+
 ## 1.4.31
 
 ### Patch Changes